backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
google-genai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
sqlalchemy
